worker/requirements.txt
psycopg2-binary
pypdf
sentence-transformers
requests
python-dotenv
supabase
langchain
langchain-text-splitters
